Removing the LazyConnectionWrapper completely
